requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
importlib_metadata==8.5.0
iniconfig==2.0.0
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg==3.2.3
PyJWT==1.7.1
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.1
tomli==2.0.2
tzdata==2024.2
wheel==0.44.0
